+++ ICD.Connect.Calendaring.Google/Converters/GoogleCalendarViewResponseConverter.cs
@@ -32,7 +32,7 @@ namespace ICD.Connect.Calendaring.Google.Converters
 					instance.Summary = reader.GetValueAsString();
 					break;
 				case ATTRIBUTE_UPDATED:
-					instance.Updated = reader.GetValueAsDateTime();
+					instance.Updated = reader.GetValueAsDateTime().ToUniversalTime();
 					break;
 				case ATTRIBUTE_TIMEZONE:
 					instance.TimeZone = reader.GetValueAsString();